+++ django_typer/tests/tests.py
@@ -11,6 +11,7 @@ from pathlib import Path
 from typing import Any, Tuple
 
 import django
+import pexpect
 import pytest
 import typer
 from click.exceptions import UsageError
@@ -62,7 +63,20 @@ def get_named_arguments(function):
     ]
 
 
-def run_command(command, *args, parse_json=True) -> Tuple[str, str]:
+def interact(command, *args, **kwargs):
+    cwd = os.getcwd()
+    try:
+        os.chdir(manage_py.parent)
+        return pexpect.spawn(
+            " ".join([sys.executable, f"./{manage_py.name}", command, *args]),
+            env=os.environ,
+            **kwargs,
+        )
+    finally:
+        os.chdir(cwd)
+
+
+def run_command(command, *args, parse_json=True, **kwargs) -> Tuple[str, str]:
     # we want to use the same test database that was created for the test suite run
     cwd = os.getcwd()
     try:
@@ -73,6 +87,7 @@ def run_command(command, *args, parse_json=True) -> Tuple[str, str]:
             capture_output=True,
             text=True,
             env=env,
+            **kwargs,
         )
 
         # Check the return code to ensure the script ran successfully
@@ -2753,3 +2768,93 @@ class TestHandleAsInit(TestCase):
         self.assertEqual(get_command("handle_as_init")(), "handle")
         self.assertEqual(get_command("handle_as_init", "subcommand")(), "subcommand")
         self.assertEqual(get_command("handle_as_init").subcommand(), "subcommand")
+
+
+class TestPromptOptions(TestCase):
+
+    def test_run_with_option_prompt(self):
+
+        cmd = interact("prompt", "--no-color", "cmd1", "bckohan")
+        cmd.expect("Password: ")
+        cmd.sendline("test_password")
+
+        result = cmd.read().decode("utf-8").strip().splitlines()[0]
+        self.assertEqual(result, "bckohan test_password")
+
+        cmd = interact("prompt", "--no-color", "cmd2", "bckohan")
+        result = cmd.read().decode("utf-8").strip().splitlines()[0]
+        self.assertEqual(result, "bckohan None")
+
+        cmd = interact("prompt", "--no-color", "cmd2", "bckohan", "-p")
+        cmd.expect("Password: ")
+        cmd.sendline("test_password2")
+        result = cmd.read().decode("utf-8").strip().splitlines()[0]
+        self.assertEqual(result, "bckohan test_password2")
+
+        cmd = interact("prompt", "--no-color", "cmd3", "bckohan")
+        result = cmd.read().decode("utf-8").strip().splitlines()[0]
+        self.assertEqual(result, "bckohan default")
+
+        cmd = interact("prompt", "--no-color", "cmd3", "bckohan", "-p")
+        cmd.expect(r"Password \[default\]: ")
+        cmd.sendline("test_password3")
+        result = cmd.read().decode("utf-8").strip().splitlines()[0]
+        self.assertEqual(result, "bckohan test_password3")
+
+        cmd = interact("prompt", "--no-color", "group1", "cmd4", "bckohan")
+        cmd.expect(r"Flag: ")
+        cmd.sendline("test_flag")
+        cmd.expect(r"Password: ")
+        cmd.sendline("test_password4")
+        result = cmd.read().decode("utf-8").strip().splitlines()[0]
+        self.assertEqual(result, "test_flag bckohan test_password4")
+
+    def test_call_with_option_prompt(self):
+        self.assertEqual(
+            call_command(
+                "prompt", "--no-color", "cmd1", "bckohan", password="test_password"
+            ),
+            "bckohan test_password",
+        )
+
+        self.assertEqual(
+            call_command("prompt", "--no-color", "cmd2", "bckohan"), "bckohan None"
+        )
+
+        self.assertEqual(
+            call_command(
+                "prompt", "--no-color", "cmd2", "bckohan", "-p", "test_password2"
+            ),
+            "bckohan test_password2",
+        )
+
+        self.assertEqual(
+            call_command("prompt", "--no-color", "cmd3", "bckohan"), "bckohan default"
+        )
+
+        self.assertEqual(
+            call_command(
+                "prompt", "--no-color", "cmd3", "bckohan", password="test_password3"
+            ),
+            "bckohan test_password3",
+        )
+
+        self.assertEqual(
+            call_command(
+                "prompt",
+                "--no-color",
+                "group1",
+                "-f",
+                "test_flag",
+                "cmd4",
+                "bckohan",
+                password="test_password4",
+            ),
+            "test_flag bckohan test_password4",
+        )
+
+        # this doesn't work!
+        # self.assertEqual(
+        #     call_command('prompt', '--no-color', 'group1', 'cmd4', 'bckohan', flag='test_flag', password='test_password4'),
+        #     'test_flag bckohan test_password4'
+        # )